--- a/lib/steep/drivers/langserver.rb
+++ b/lib/steep/drivers/langserver.rb
@@ -43,7 +43,8 @@ module Steep
           reader: reader,
           writer: writer,
           interaction_worker: interaction_worker,
-          typecheck_workers: typecheck_workers
+          typecheck_workers: typecheck_workers,
+          refork: Steep.can_fork?,
         )
         master.typecheck_automatically = true