--- a/lib/steep/lsp_patch.rb
+++ b/lib/steep/lsp_patch.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
+# https://github.com/mtsmfm/language_server-protocol-ruby/pull/112
 module LanguageServer
   module Protocol
     module Transport
@@ -7,13 +8,13 @@ module LanguageServer
         class Reader
           def close
             @io.close
-          end
+          end unless method_defined?(:close)
         end
 
         class Writer
           def close
             @io.close
-          end
+          end unless method_defined?(:close)
         end
       end
     end