--- a/lib/steep/server/custom_methods.rb
+++ b/lib/steep/server/custom_methods.rb
@@ -72,6 +72,18 @@ module Steep
           { id: id, result: result }
         end
       end
+
+      module Refork
+        METHOD = "$/steep/refork"
+
+        def self.request(id, params)
+          { method: METHOD, id: id, params: params }
+        end
+
+        def self.response(id, result)
+          { id: id, result: result }
+        end
+      end
     end
   end
 end