--- a/lib/steep/server/type_check_worker.rb
+++ b/lib/steep/server/type_check_worker.rb
@@ -51,7 +51,7 @@ module Steep
 
       include ChangeBuffer
 
-      def initialize(project:, reader:, writer:, assignment:, commandline_args:, io_socket:, buffered_changes: nil)
+      def initialize(project:, reader:, writer:, assignment:, commandline_args:, io_socket:, buffered_changes: nil, service: nil)
         super(project: project, reader: reader, writer: writer)
 
         @assignment = assignment
@@ -61,6 +61,7 @@ module Steep
         @commandline_args = commandline_args
         @current_type_check_guid = nil
         @io_socket = io_socket
+        @service = service
 
         if io_socket
           Signal.trap "SIGCHLD" do
@@ -130,9 +131,15 @@ module Steep
             reader = LanguageServer::Protocol::Transport::Io::Reader.new(stdin)
             writer = LanguageServer::Protocol::Transport::Io::Writer.new(stdout)
             Steep.logger.info("Reforked worker: #{Process.pid}, params: #{request[:params]}")
-            assignment = Services::PathAssignment.new(max_index: request[:params][:max_index], index: request[:params][:index])
+            index = request[:params][:index]
+            assignment = Services::PathAssignment.new(max_index: request[:params][:max_index], index: index)
 
-            self.class.new(project: project, reader: reader, writer: writer, assignment: assignment, commandline_args: commandline_args, io_socket: nil, buffered_changes: buffered_changes).run()
+            worker = self.class.new(project: project, reader: reader, writer: writer, assignment: assignment, commandline_args: commandline_args, io_socket: nil, buffered_changes: buffered_changes, service: service)
+
+            tags = Steep.logger.formatter.current_tags.dup
+            tags[tags.find_index("typecheck:typecheck@0")] = "typecheck:typecheck@#{index}-reforked"
+            Steep.logger.formatter.push_tags(tags)
+            worker.run()
 
             raise "unreachable"
           end