--- a/rbs_collection.steep.yaml
+++ b/rbs_collection.steep.yaml
@@ -35,3 +35,4 @@ gems:
   - name: optparse
   - name: digest
   - name: uri
+  - name: socket