--- a/sig/steep/drivers/langserver.rbs
+++ b/sig/steep/drivers/langserver.rbs
@@ -15,6 +15,8 @@ module Steep
 
       attr_reader jobs_option: Utils::JobsOption
 
+      attr_accessor refork: bool
+
       include Utils::DriverHelper
 
       def initialize: (stdout: untyped, stderr: untyped, stdin: untyped) -> void