… Gemfile.lock
… rbs_collection.steep.lock.yaml
… sig/steep/server/custom_methods.rbs
… sig/steep/server/type_check_worker.rbs
… sig/steep/server/worker_process.rbs
… steep.gemspec