@soutaro 2025-02-28
Thanks @pocke ! Great work!! 👏 

I'm wondering if it's necessary to fork type check workers from the main process at the first time, not forking from the primary worker from the first time. It seems like there is (almost) no benefit to do so because the memory shared cannot be increased. So, wondering if changing the architecture to fork from the primary process from the first time makes the code simpler.