@soutaro 2025-03-03
Talked with @pocke about reforking from the first type checking. The reforking aims to reuse CoW memory by forked after type checking session from the primary type check worker, so reforking from the cold primary worker doesn't make sense.